Make get all articles work
