--- pkg/handlers/AddArticle.go
@@ -1,1 +1,0 @@
-func (h handler) AddArticle(w http.ResponseWriter, r *http.Request) { ...}
