--- pkg/handlers/GetArticle.go
@@ -1,1 +1,0 @@
-func (h handler) GetArticle(w http.ResponseWriter, r *http.Request) { ...
