--- pkg/handlers/UpdateArticle.go
@@ -1,1 +1,0 @@
-func (h handler) UpdateArticle(w http.ResponseWriter, r *http.Request) { ...
